Get and Save File Information

Get and Save File Information

get_and_save_file_info(
  root_dir,
  output_dir,
  check_dirs = TRUE,
  format = "%Y-%m-%d_%H%M",
  ...
)

Arguments

root_dir

path to the directory from which to start searching for files

output_dir

path to the output directory. In this directory, a file "path-info_<date-time>_<parent-folder>.csv" will be generated with <date-time> being a date and time string in yyyy-mm-dd_HHMM format and <parent-folder> being the last path segment of root_dir

check_dirs

if TRUE (default) it is checked in advance if both root_dir and output_dir exist. Switch this off if e.g. network paths are wrongly considered to be non-existing.

format

format string specifying how to format the part of the filename intended to contain date (and, if required, time) information. Default: "%Y-%m-%d_%H%M"

...

further arguments passed to get_recursive_file_info and finally to fs::dir_info. Set e.g. fail = FALSE to avoid failure due to insufficient access permissions.

Value

full path to the file to which all file information were written