Call file.info recursively on files below a root folder

Call file.info recursively on files below a root folder

get_recursive_file_info(root_dir, pattern = NULL, all = TRUE, ..., dbg = TRUE)

Arguments

root_dir
path to the root directory from which to start the recursive search for files

pattern
regular expression matching the names of the files to be considered. By default, all files are considered.

all
if TRUE (default) hidden files are also returned

...
further arguments passed to fs::dir_info

dbg
if TRUE (default) progress messages are shown