Prepare and Filter Path Data for Treemap Plot

Prepare and Filter Path Data for Treemap Plot

prepare_for_treemap(path_data, root_path = "", variable = "size", ...)

Arguments

path_data

data frame as returned by read_file_info

root_path

path to the folder that contains all paths to be considered. By setting the root path to "/path/to/root" you can "zoom into" the treeplot, showing the contents below "/path/to/root" only. If root_path is "" (default) all paths in path_data are considered.

variable

name(s) of variable(s) to be selected. Default: "size"

...

further arguments passed to remove_common_root, such as n_keep (number of last segments to be kept from the common first part of all paths)