Read Paths From File

Read paths from a file that contains one path per line. Convert backslashes to slashes and sort the paths (by default)

read_paths_(
  file,
  fileEncoding = NULL,
  encoding = NULL,
  ...,
  do_sort = TRUE,
  expected_encodings = c("windows-1252", "ISO-8859-1")
)

Arguments

file

full path to the text file containing paths

fileEncoding

The name of the encoding to be assumed. Passed as encoding to file, see there.

encoding

passed to readLines.

...

arguments passed to readLinesWithEncoding, such as n

do_sort

if TRUE (default), the vector of paths is sorted alphanumerically.

expected_encodings

vector of names of file encodings that are expected to occur. Expected encodings will be preferred.

Value

vector of character